Empezando con ejercicio 15/03

+++ src/main/java/Repaso_Fecha_5_27_2018/PruebaPreviaBarbutoo_11_12/Queue.java
@@ -27,7 +27,7 @@ class Queue  {
     //--------------------------------------------------------------
     public int remove()         // take item from front of queue
     {
-        long temp = queArray[front++]; // get value and incr front
+        int temp = queArray[front++]; // get value and incr front
         if(front == maxSize)           // deal with wraparound
             front = 0;
         nItems--;                      // one less item
